+++ src/import.ts
@@ -1,14 +1,20 @@
-import * as admin from "firebase-admin";
-import * as fs from "fs";
+import * as admin from 'firebase-admin';
+import * as fs from 'fs';
+
+export interface IImportOptions {
+  dates?: string[];
+  geos?: string[];
+  refs?: string[];
+}
 
 /**
  * Convert time array in document object
- * @param dateArray 
- * @param documentObj 
+ * @param dateArray
+ * @param documentObj
  */
 const updateTime = (dateArray, documentObj): void => {
   dateArray.forEach(c => {
-    c.split(".").reduce((acc, cur, i, a) => {
+    c.split('.').reduce((acc, cur, i, a) => {
       if (!a[i + 1] && acc[cur] && acc[cur]._seconds) {
         acc[cur] = new Date(acc[cur]._seconds * 1000);
       } else return acc[cur] || {};
@@ -20,32 +26,30 @@ const updateTime = (dateArray, documentObj): void => {
  * Restore data to firestore
  *
  * @param {string} fileName
- * @param {Array<string>} dateArray
- * @param {Array<string>} geoArray
+ * @param {IImportOptions} options
  */
 export const restore = (
   fileName: string,
-  dateArray: Array<string>,
-  geoArray: Array<string>
+  options: IImportOptions
 ): Promise<any> => {
   const db = admin.firestore();
 
   return new Promise((resolve, reject) => {
-    if (typeof fileName === "object") {
+    if (typeof fileName === 'object') {
       let dataArray = fileName;
 
-      updateCollection(db, dataArray, dateArray, geoArray)
+      updateCollection(db, dataArray, options)
         .then(() => {
           resolve({
             status: true,
-            message: "Collection successfully imported!"
+            message: 'Collection successfully imported!'
           });
         })
         .catch(error => {
           reject({ status: false, message: error.message });
         });
     } else {
-      fs.readFile(fileName, "utf8", function(err, data) {
+      fs.readFile(fileName, 'utf8', function(err, data) {
         if (err) {
           console.log(err);
           reject({ status: false, message: err.message });
@@ -54,11 +58,11 @@ export const restore = (
         // Turn string from file to an Array
         let dataArray = JSON.parse(data);
 
-        updateCollection(db, dataArray, dateArray, geoArray)
+        updateCollection(db, dataArray, options)
           .then(() => {
             resolve({
               status: true,
-              message: "Collection successfully imported!"
+              message: 'Collection successfully imported!'
             });
           })
           .catch(error => {
@@ -80,33 +84,30 @@ export const restore = (
 const updateCollection = async (
   db,
   dataArray: Array<any>,
-  dateArray: Array<string>,
-  geoArray: Array<string>
+  options: IImportOptions
 ) => {
   for (var index in dataArray) {
     var collectionName = index;
     for (var doc in dataArray[index]) {
       if (dataArray[index].hasOwnProperty(doc)) {
-        if (dataArray[index][doc]["subCollection"]) {
-          const subCollections = dataArray[index][doc]["subCollection"];
-          delete dataArray[index][doc]["subCollection"];
+        if (dataArray[index][doc]['subCollection']) {
+          const subCollections = dataArray[index][doc]['subCollection'];
+          delete dataArray[index][doc]['subCollection'];
           await startUpdating(
             db,
             collectionName,
             doc,
             dataArray[index][doc],
-            dateArray,
-            geoArray
+            options
           );
-          await updateCollection(db, subCollections, [], []);
+          await updateCollection(db, subCollections, {});
         } else {
           await startUpdating(
             db,
             collectionName,
             doc,
             dataArray[index][doc],
-            dateArray,
-            geoArray
+            options
           );
         }
       }
@@ -128,25 +129,34 @@ const startUpdating = (
   collectionName: string,
   doc: string,
   data: object,
-  dateArray: Array<string>,
-  geoArray: Array<string>
+  options: IImportOptions
 ) => {
   let parameterValid = true;
 
-  if (typeof dateArray === "object" && dateArray.length > 0) {
-    updateTime(dateArray, data);    
+  // Update date value
+  if (options.dates && options.dates.length > 0) {
+    updateTime(options.dates, data);
+  }
+
+  // reference key
+  if (options.refs && options.refs.length > 0) {
+    options.refs.forEach(ref => {
+      if (data.hasOwnProperty(ref)) {
+        data[ref] = db.doc(data[ref]);
+      }
+    });
   }
 
   // Enter geo value
-  if (typeof geoArray !== "undefined" && geoArray.length > 0) {
-    geoArray.forEach(geo => {
+  if (options.geos && options.geos.length > 0) {
+    options.geos.forEach(geo => {
       if (data.hasOwnProperty(geo)) {
         data[geo] = new admin.firestore.GeoPoint(
           data[geo]._latitude,
           data[geo]._longitude
         );
       } else {
-        console.log("Please check your geo parameters!!!", geoArray);
+        console.log('Please check your geo parameters!!!', options.geos);
         parameterValid = false;
       }
     });
@@ -159,7 +169,7 @@ const startUpdating = (
         .set(data)
         .then(() => {
           console.log(`${doc} was successfully added to firestore!`);
-          resolve("Data written!");
+          resolve('Data written!');
         })
         .catch(error => {
           console.log(error);