+++ CHANGELOG.md
@@ -16,6 +16,15 @@
 #### - :nail_care: [Polish]
 
 - Update new packages
+- Check app before initializing
+
+#### - :boom: [Breaking Change]
+
+- Changed options for date & location & ref
+
+#### - :rocket: [New Feature]
+
+- Added support reference type
 
 ## [0.3.2] - 2020-01-03